move TopIntent() method into RecognizerResult so any recognizer result has TopIntent()
